Refactor cgen UI code
