+++ modeler/cayenne-modeler/src/main/java/org/apache/cayenne/modeler/editor/cgen/GeneratorController.java
@@ -20,28 +20,16 @@
 package org.apache.cayenne.modeler.editor.cgen;
 
 import org.apache.cayenne.gen.CgenConfiguration;
-import org.apache.cayenne.map.Embeddable;
-import org.apache.cayenne.map.EmbeddableAttribute;
-import org.apache.cayenne.map.EmbeddedAttribute;
-import org.apache.cayenne.map.ObjAttribute;
-import org.apache.cayenne.map.ObjEntity;
-import org.apache.cayenne.map.ObjRelationship;
 import org.apache.cayenne.modeler.Application;
 import org.apache.cayenne.modeler.pref.FSPath;
 import org.apache.cayenne.modeler.util.CayenneController;
-import org.apache.cayenne.modeler.util.CodeValidationUtil;
 import org.apache.cayenne.modeler.util.TextAdapter;
 import org.apache.cayenne.swing.BindingBuilder;
 import org.apache.cayenne.util.Util;
-import org.apache.cayenne.validation.BeanValidationFailure;
-import org.apache.cayenne.validation.SimpleValidationFailure;
-import org.apache.cayenne.validation.ValidationFailure;
-import org.apache.cayenne.validation.ValidationResult;
 
 import javax.swing.JButton;
 import javax.swing.JFileChooser;
 import java.io.File;
-import java.util.function.Predicate;
 
 /**
  * @since 4.1
@@ -82,214 +70,6 @@ public abstract class GeneratorController extends CayenneController {
     }
 
     public abstract void updateConfiguration(CgenConfiguration cgenConfiguration);
-
-    public void validateEmbeddable(ValidationResult validationBuffer, Embeddable embeddable) {
-        ValidationFailure embeddableFailure = validateEmbeddable(embeddable);
-        if (embeddableFailure != null) {
-            validationBuffer.addFailure(embeddableFailure);
-            return;
-        }
-
-        for (EmbeddableAttribute attribute : embeddable.getAttributes()) {
-            ValidationFailure failure = validateEmbeddableAttribute(attribute);
-            if (failure != null) {
-                validationBuffer.addFailure(failure);
-                return;
-            }
-        }
-    }
-
-    private ValidationFailure validateEmbeddableAttribute(EmbeddableAttribute attribute) {
-        return validateAttribute(attribute.getEmbeddable().getClassName(), attribute.getName(), attribute.getType());
-    }
-
-    private ValidationFailure validateAttribute(String objectName, String attributeName, String attributeType) {
-        ValidationFailure emptyName = BeanValidationFailure.validateNotEmpty(objectName, "attribute.name",
-                attributeName);
-        if (emptyName != null) {
-            return emptyName;
-        }
-
-        ValidationFailure badName = CodeValidationUtil.validateJavaIdentifier(objectName, "attribute.name",
-                attributeName);
-        if (badName != null) {
-            return badName;
-        }
-
-        ValidationFailure emptyType = BeanValidationFailure.validateNotEmpty(objectName, "attribute.type",
-                attributeType);
-        if (emptyType != null) {
-            return emptyType;
-        }
-
-        return BeanValidationFailure.validateJavaClassName(objectName, "attribute.type", attributeType);
-    }
-
-    private ValidationFailure validateEmbeddable(Embeddable embeddable) {
-
-        String name = embeddable.getClassName();
-
-        ValidationFailure emptyClass = BeanValidationFailure.validateNotEmpty(name, "className",
-                embeddable.getClassName());
-        if (emptyClass != null) {
-            return emptyClass;
-        }
-
-        return BeanValidationFailure.validateJavaClassName(name, "className", embeddable.getClassName());
-    }
-
-    public void validateEntity(ValidationResult validationBuffer, ObjEntity entity, boolean clientValidation) {
-
-        ValidationFailure entityFailure = validateEntity(clientValidation ? entity.getClientEntity() : entity);
-        if (entityFailure != null) {
-            validationBuffer.addFailure(entityFailure);
-            return;
-        }
-
-        for (ObjAttribute attribute : entity.getAttributes()) {
-            if (attribute instanceof EmbeddedAttribute) {
-                EmbeddedAttribute embeddedAttribute = (EmbeddedAttribute) attribute;
-                for (ObjAttribute subAttribute : embeddedAttribute.getAttributes()) {
-                    ValidationFailure failure = validateEmbeddedAttribute(subAttribute);
-                    if (failure != null) {
-                        validationBuffer.addFailure(failure);
-                        return;
-                    }
-                }
-            } else {
-
-                ValidationFailure failure = validateAttribute(attribute);
-                if (failure != null) {
-                    validationBuffer.addFailure(failure);
-                    return;
-                }
-            }
-        }
-
-        for (ObjRelationship rel : entity.getRelationships()) {
-            ValidationFailure failure = validateRelationship(rel, clientValidation);
-            if (failure != null) {
-                validationBuffer.addFailure(failure);
-                return;
-            }
-        }
-    }
-
-    private ValidationFailure validateEntity(ObjEntity entity) {
-
-        String name = entity.getName();
-
-        if (entity.isGeneric()) {
-            return new SimpleValidationFailure(name, "Generic class");
-        }
-
-        ValidationFailure emptyClass = BeanValidationFailure.validateNotEmpty(name, "className", entity.getClassName());
-        if (emptyClass != null) {
-            return emptyClass;
-        }
-
-        ValidationFailure badClass = BeanValidationFailure.validateJavaClassName(name, "className",
-                entity.getClassName());
-        if (badClass != null) {
-            return badClass;
-        }
-
-        if (entity.getSuperClassName() != null) {
-            return BeanValidationFailure.validateJavaClassName(name, "superClassName", entity.getSuperClassName());
-        }
-
-        return null;
-    }
-
-    private ValidationFailure validateAttribute(ObjAttribute attribute) {
-        return validateAttribute(attribute.getEntity().getName(), attribute.getName(), attribute.getType());
-    }
-
-    private ValidationFailure validateEmbeddedAttribute(ObjAttribute attribute) {
-
-        String name = attribute.getEntity().getName();
-
-        // validate embeddedAttribute and attribute names
-        // embeddedAttribute returned attibute as
-        // [name_embeddedAttribute].[name_attribute]
-        String[] attributes = attribute.getName().split("\\.");
-        String nameEmbeddedAttribute = attributes[0];
-        int beginIndex = attributes[0].length();
-        String attr = attribute.getName().substring(beginIndex + 1);
-
-        ValidationFailure emptyEmbeddedName = BeanValidationFailure.validateNotEmpty(name, "attribute.name",
-                nameEmbeddedAttribute);
-        if (emptyEmbeddedName != null) {
-            return emptyEmbeddedName;
-        }
-
-        ValidationFailure badEmbeddedName = CodeValidationUtil.validateJavaIdentifier(name, "attribute.name",
-                nameEmbeddedAttribute);
-        if (badEmbeddedName != null) {
-            return badEmbeddedName;
-        }
-
-        return validateAttribute(name, attr, attribute.getType());
-    }
-
-    private ValidationFailure validateRelationship(ObjRelationship relationship, boolean clientValidation) {
-
-        String name = relationship.getSourceEntity().getName();
-
-        ValidationFailure emptyName = BeanValidationFailure.validateNotEmpty(name, "relationship.name",
-                relationship.getName());
-        if (emptyName != null) {
-            return emptyName;
-        }
-
-        ValidationFailure badName = CodeValidationUtil.validateJavaIdentifier(name, "relationship.name",
-                relationship.getName());
-        if (badName != null) {
-            return badName;
-        }
-
-        if (!relationship.isToMany()) {
-
-            ObjEntity targetEntity = relationship.getTargetEntity();
-
-            if (clientValidation && targetEntity != null) {
-                targetEntity = targetEntity.getClientEntity();
-            }
-
-            if (targetEntity == null) {
-
-                return new BeanValidationFailure(name, "relationship.targetEntity", "No target entity");
-            } else if (!targetEntity.isGeneric()) {
-                ValidationFailure emptyClass = BeanValidationFailure.validateNotEmpty(name,
-                        "relationship.targetEntity.className", targetEntity.getClassName());
-                if (emptyClass != null) {
-                    return emptyClass;
-                }
-
-                return BeanValidationFailure.validateJavaClassName(name, "relationship.targetEntity.className",
-                        targetEntity.getClassName());
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Returns a predicate for default entity selection in a given mode.
-     */
-    public Predicate<Object> getDefaultClassFilter() {
-        return object -> {
-            if (object instanceof ObjEntity) {
-                return getParentController().getProblem(((ObjEntity) object).getName()) == null;
-            }
-
-            if (object instanceof Embeddable) {
-                return getParentController().getProblem(((Embeddable) object).getClassName()) == null;
-            }
-
-            return false;
-        };
-    }
 
     /**
      * An action method that pops up a file chooser dialog to pick the
